+++ mrcrowbar/encoding.py
@@ -1,7 +1,110 @@
 import codecs
 import re
 import struct
+import logging
+logger = logging.getLogger( __name__ )
 
+# Python doesn't provide a programmatic way of fetching the supported codec list.
+# The below list is taken from the 3.7 manual.
+CODECS = [
+    'ascii',
+    'big5',
+    'big5hkscs',
+    'cp037',
+    'cp273',
+    'cp424',
+    'cp437',
+    'cp500',
+    'cp720',
+    'cp737',
+    'cp775',
+    'cp850',
+    'cp852',
+    'cp855',
+    'cp856',
+    'cp857',
+    'cp858',
+    'cp860',
+    'cp861',
+    'cp862',
+    'cp863',
+    'cp864',
+    'cp865',
+    'cp866',
+    'cp869',
+    'cp874',
+    'cp875',
+    'cp932',
+    'cp949',
+    'cp950',
+    'cp1006',
+    'cp1026',
+    'cp1125',
+    'cp1140',
+    'cp1250',
+    'cp1251',
+    'cp1252',
+    'cp1253',
+    'cp1254',
+    'cp1255',
+    'cp1256',
+    'cp1257',
+    'cp1258',
+    'euc_jp',
+    'euc_jis_2004',
+    'euc_jisx0213',
+    'euc_kr',
+    'gb2312',
+    'gbk',
+    'gb18030',
+    'hz',
+    'iso2022_jp',
+    'iso2022_jp_1',
+    'iso2022_jp_2',
+    'iso2022_jp_2004',
+    'iso2022_jp_3',
+    'iso2022_jp_ext',
+    'iso2022_kr',
+    'latin_1',
+    'iso8859_2',
+    'iso8859_3',
+    'iso8859_4',
+    'iso8859_5',
+    'iso8859_6',
+    'iso8859_7',
+    'iso8859_8',
+    'iso8859_9',
+    'iso8859_10',
+    'iso8859_11',
+    'iso8859_13',
+    'iso8859_14',
+    'iso8859_15',
+    'iso8859_16',
+    'johab',
+    'koi8_r',
+    'koi8_t',
+    'koi8_u',
+    'kz1048',
+    'mac_cyrillic',
+    'mac_greek',
+    'mac_iceland',
+    'mac_latin2',
+    'mac_roman',
+    'mac_turkish',
+    'ptcp154',
+    'shift_jis',
+    'shift_jis_2004',
+    'shift_jisx0213',
+    'utf_32',
+    'utf_32_be',
+    'utf_32_le',
+    'utf_16',
+    'utf_16_be',
+    'utf_16_le',
+    'utf_7',
+    'utf_8',
+    'utf_8_sig'
+]
 
 REGEX_CHARS = """()[]{}?*+-|^$\\.&~#="""
 byte_escape = lambda char: '\\x{:02x}'.format( char ).encode( 'utf8' )
@@ -61,6 +164,28 @@ def regex_pattern_to_bytes( pattern, encoding='utf8', fixed_string=False, hex_fo
             pointer += 1
     return bytes( result )
 
+
+def regex_unknown_encoding_match( string, char_size=1 ):
+    match_map = {}
+    pattern = bytearray()
+    for i, char in enumerate( string ):
+        if char not in match_map:
+            match_id = len( match_map )
+            match_group = '?P<p{}>.'.format( match_id ).encode( 'utf8' )
+            if char_size != 1:
+                match_group += b'{' + '{}'.format( char_size ).encode( 'utf8' ) + b'}'
+            if len( pattern ) == 0:
+                pattern += b'(' + match_group + b')'
+            else:
+                pattern += b'(' + match_group + b'(?<!'
+                pattern += b'|'.join( ['(?P=p{})'.format( match_map[c] ).encode( 'utf8' ) for c in match_map if c != char] )
+                pattern += b'))'
+            match_map[char] = match_id
+        else:
+            pattern += '(?P=p{})'.format( match_map[char] ).encode( 'utf8' )
+    if len( string ) == len( match_map ):
+        logger.warning( 'Input has no repeated characters! This can make an enormous number of false matches, and is likely not what you want' )
+    return match_map, bytes( pattern )
 
 
 RAW_TYPE_NAME = {