+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Unreleased
 ==========
 - common.bounds: Allow negative indexing for hexdumps etc.
 - blocks.Block: Fall back to Unknown for child Blocks that fail to parse
+- utils: Add find, find_encoded and find_unknown functions
 
 0.8.0 - 2020-07-03
 ==================